requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
statsmodels
pmdarima
xgboost
lightgbm
yfinance
jupyter
notebook
ipywidgets
joblib
tqdm
openpyxl
python-pptx
fpdf
